Fix multi line synthesis
The synthesis route was blindly calling .replace on text without making sure it was a string first. If you tried to use the Multi Line method on the synthesis page text was a list type instead of a string and the folder name creation would throw an exception. The single line and paragraph methods correctly typed text as a string.